requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
django-environ
django-extensions
drf-yasg
djangorestframework-simplejwt
django-allauth
dj-rest-auth
psycopg2-binary
Pillow
django-storages
python-dotenv
gunicorn
whitenoise
requests
black
isort
flake8
ipython
django-unfold